refactor(plugins): move BasicStatsCalculator into basic-stats plugin slice
Relocate the class from src/shared/stats/basic/ to
src/plugins/basic-stats/application/ per the plugin-system design (D9):
match-resume and player-stats stay in core because they have other
consumers, but basic-stats itself has none besides this class.

Only intra-slice relative import paths change; behavior is unchanged.
src/index.ts still constructs and subscribes BasicStatsCalculator
directly (bootstrapPlugins wiring lands in a later PR) and only needs
its import path updated to follow the move.
