+++ src/plugins/basic-stats/application/BasicStatsCalculator.test.ts
